+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - Muuuuch better tests
 - Embedded permissions object cannot be overwritten
 - When a document has embedded permissions, those permissions will be checks when a save or remove is being done. That way someone cannot write to an object in a way that changes their permssions and then try to save it.
+- Does not add permissions when a doc is empty